Move infusion helper logic into redux using reselect
