+++ src/utils/gw2-data.js
@@ -925,6 +925,77 @@ export const INFUSIONS = [
   { id: 37136, attribute: 'Vitality' }, // vitality
 ];
 
+export const infusionIds = {
+  '+1 agony': { id: 49424, cost: 7 },
+  '+2 agony': { id: 49425, cost: 164 },
+  '+3 agony': { id: 49426, cost: 478 },
+  '+4 agony': { id: 49427, cost: 1106 },
+  '+5 agony': { id: 49428, cost: 2632 },
+  '+6 agony': { id: 49429, cost: 4878 },
+  '+7 agony': { id: 49430 },
+  '+8 agony': { id: 49431 },
+  '+9 agony': { id: 49432 },
+  '+10 agony': { id: 49433 },
+  '+11 agony': { id: 49434 },
+  '+12 agony': { id: 49435 },
+  '+13 agony': { id: 49436 },
+  '+14 agony': { id: 49437 },
+  '+15 agony': { id: 49438 },
+  '+16 agony': { id: 49439 },
+  '+17 agony': { id: 49440 },
+  '+18 agony': { id: 49441 },
+  '+19 agony': { id: 49442 },
+  '+20 agony': { id: 49443 },
+  '+21 agony': { id: 49444 },
+  '+22 agony': { id: 49445 },
+  '+23 agony': { id: 49446 },
+  '+24 agony': { id: 49447 },
+
+  '+9 stat': {
+    'Power': { id: 37131 },
+    'Precision': { id: 37132 },
+    'Condition Damage': { id: 37130 },
+    'Expertise': { id: 86113 },
+    'Concentration': { id: 86180 },
+    'Healing Power': { id: 37125 },
+    'Toughness': { id: 37135 },
+    'Vitality': { id: 37136 },
+  },
+
+  '+7 stat': {
+    'Power': { id: 37127 },
+    'Precision': { id: 37128 },
+    'Condition Damage': { id: 37129 },
+    'Expertise': { id: 86150 },
+    'Concentration': { id: 85881 },
+    'Healing Power': { id: 37123 },
+    'Toughness': { id: 37133 },
+    'Vitality': { id: 37134 },
+  },
+
+  '+5 stat': {
+    'Power': { id: 39620 },
+    'Precision': { id: 39621 },
+    'Condition Damage': { id: 39619 },
+    'Expertise': { id: 85971 },
+    'Concentration': { id: 86338 },
+    'Healing Power': { id: 39616 },
+    'Toughness': { id: 39617 },
+    'Vitality': { id: 39618 },
+  },
+
+  'WvW stat': {
+    'Power': { id: 43254 },
+    'Precision': { id: 43255 },
+    'Condition Damage': { id: 43253 },
+    'Expertise': { id: 87218 },
+    'Concentration': { id: 86986 },
+    'Healing Power': { id: 43250 },
+    'Toughness': { id: 43251 },
+    'Vitality': { id: 43252 },
+  },
+};
+
 export const ARMOR_IDS = {
   // Helm, Shoulders, Chest, Hands, Leggings, Feet
   HEAVY: [86198, 85829, 85814, 85949, 86079, 86237],
